Create a TheCeramicText.vue component, and move the corresponding html and js there
